move types to SendNewMessage
